Makefile : clean removes bonus objs & Fix .cc rule
by adding both the .o and .cc file to the target of the rule, Make knows that this rulw will create both files.
Added the bonus objects to be deleted unpon clean.
The removal of compile_commands.json has been moved to fclean, as it is a useful file on it's own, and not just an intermediary file